add admin simple verification

+++ database/crud/db_users.py
@@ -12,6 +12,8 @@ def validate_user_id(current_user: ActiveUser, user_id: int):
     validate_id = user_id
     if current_user_id == validate_id:
         return True
+    elif current_user_id == 1:
+        return True
     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
                         detail="You don't have access to change other Users credentials. "
                                "Make sure to use registered ID of you own.",
@@ -49,6 +51,10 @@ def get_user(user_id: int | None, login: str | None, db: Session) -> Type[DbUser
 
 def update_user(user_id: int | None, db: Session, request: UpdateUser) -> Type[DbUsers] | bool:
     """Find record in DbUsers with given user_id or login. Update record with new data."""
+    if user_id == 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Admin can't be modified after creation",
+                            )
     new_login = request.new_login.lower().strip().replace(" ", "")
     new_hash_password = Hash().bcrypt_pass(request.new_password.strip().replace(" ", ""))
     user = get_user(login=None, user_id=user_id, db=db)
@@ -59,24 +65,18 @@ def update_user(user_id: int | None, db: Session, request: UpdateUser) -> Type[D
         DbUsers.password: new_hash_password,
     })
     db.commit()
-    db.refresh(user)
     return user
 
 
-def delete_user(user_id: int | None, login: str | None, db: Session) -> bool:
+def delete_user(user_id: int | None, db: Session) -> bool:
     """Delete record from DbUsers with given user_id or login."""
-    if login:
-        del_login = login.lower()
-        user = get_user(login=del_login, user_id=None, db=db)
-        if user is None:
-            return False
-        db.delete(user)
-        db.commit()
-        return True
+    if user_id == 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Admin Id can't be deleted",
+                            )
     user = get_user(login=None, user_id=user_id, db=db)
     if user is None:
         return False
     db.delete(user)
     db.commit()
     return True
-
